cleanup to wasm test suite remove docker and other fixes
- all the docker stuff is now gone, the only requirement is a local install of Chrome
- test programs moved into tests/wasm/testdata and "_pgm" removed from the name, unnecessary tinygo build tags removed
- all of the test setup stuff now lives in one file, setup_test.go, and it runs the necessary http server internally
- a temp dir is used for all output files
- any remaining `must` calls replaced with `mustf` and explanations added
- added t.Parallel(), seems to work fine, can be easily removed later if there are memory problems with too many open headless Chrome instances

testing fmtprint against all versions again in the pipeline to see if i can debug
