--- tests/wasm/0dockerfile
@@ -1,10 +1,0 @@
-
-FROM chromedp/headless-shell:latest
-
-RUN apt-get update && apt-get install dumb-init
-
-COPY server /bin/server
-COPY 0docker_entrypoint.sh /0docker_entrypoint.sh
-RUN chmod 755 /bin/server /0docker_entrypoint.sh
-
-ENTRYPOINT ["/usr/bin/dumb-init", "--", "/0docker_entrypoint.sh"]
