+++ tests/wasm/fmtprint_test.go
@@ -1,5 +1,3 @@
-// +build go1.14
-
 package wasm
 
 import (
@@ -11,7 +9,9 @@ import (
 
 func TestFmtprint(t *testing.T) {
 
-	err := run("tinygo build -o fmtprint_pgm.wasm -target wasm fmtprint_pgm.go")
+	t.Parallel()
+
+	err := run("tinygo build -o " + wasmTmpDir + "/fmtprint.wasm -target wasm testdata/fmtprint.go")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -21,13 +21,13 @@ func TestFmtprint(t *testing.T) {
 
 	var log1 string
 	err = chromedp.Run(ctx,
-		chromedp.Navigate("http://localhost:8826/run?file=fmtprint_pgm.wasm"),
+		chromedp.Navigate("http://localhost:8826/run?file=fmtprint.wasm"),
 		chromedp.Sleep(time.Second),
 		chromedp.InnerHTML("#log", &log1),
-		waitLog(`test from fmtprint_pgm 1
-test from fmtprint_pgm 2
-test from fmtprint_pgm 3
-test from fmtprint_pgm 4`),
+		waitLog(`test from fmtprint 1
+test from fmtprint 2
+test from fmtprint 3
+test from fmtprint 4`),
 	)
 	t.Logf("log1: %s", log1)
 	if err != nil {
