+++ tests/wasm/testdata/chan.go
@@ -1,5 +1,3 @@
-// +build tinygo
-
 package main
 
 func main() {
