+++ tests/wasm/testdata/fmtprint.go
@@ -1,0 +1,11 @@
+package main
+
+import "fmt"
+
+func main() {
+	fmt.Println("test from fmtprint 1")
+	fmt.Print("test from fmtprint 2\n")
+	fmt.Print("test from fmtp")
+	fmt.Print("rint 3\n")
+	fmt.Printf("test from fmtprint %d\n", 4)
+}
